edit template
