--- src/app/templates/edit/[id].tsx
@@ -1,3 +1,0 @@
-export default function Page() {
-  return <h1> Edit Page</h1>;
-}
