+++ src/app/templates/schema.ts
